+++ example/src/ajax/index.tsx
@@ -6,6 +6,7 @@ import Download from './download/box';
 import Crypto from './crypto/box';
 import Signature from './signature/box';
 import ErrorRequest from './error/box';
+import BigPipe from './bigpipe/box';
 import './index.css';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
@@ -36,6 +37,7 @@ export default class Ajax extends React.PureComponent {
                             <Crypto />
                             <Signature />
                             <ErrorRequest />
+                            <BigPipe />
                         </div>
                     </Page>
                     <PropsTable title="Options" of={optionsDoc} />